Split useCastSnapshotVote out of useCastVote
